LayoutTests/imported/w3c/web-platform-tests/css/cssom/CSSStyleSheet-constructable-replace-thenable-expected.txt


PASS CSSStyleSheet.replace/replaceSync() when stylesheet is a thenable
